requirements.txt
tensorflow==1.3.0
Flask_BasicAuth==0.2.0
nltk==3.2.4
colorama==0.3.9
Flask==1.0.2
numpy==1.14.5
Werkzeug==0.14.1
hyperopt==0.1
tqdm==4.23.4
pefile==2018.8.8
scipy==0.19.1
Pillow==5.2.0
protobuf==3.6.1
scikit_learn==0.19.2
typing==3.6.6
